+++ sickrage/providers/torrent/bitcannon.py
@@ -23,6 +23,7 @@ from urllib import quote_plus
 
 import sickrage
 from sickrage.core.caches.tv_cache import TVCache
+from sickrage.core.helpers import try_int
 from sickrage.providers import TorrentProvider
 
 
@@ -45,7 +46,6 @@ class BitCannonProvider(TorrentProvider):
 
     def search(self, search_strings, search_mode='eponly', epcount=0, age=0, epObj=None):
         results = []
-        items = {'Season': [], 'Episode': [], 'RSS': []}
 
         try:
             trackers = sickrage.srCore.srWebSession.get(self.urls['trackers'], cache=False).json().get('Trackers')
@@ -93,16 +93,16 @@ class BitCannonProvider(TorrentProvider):
                                                                                [quote_plus(x.encode('utf-8')) for x
                                                                                 in trackers]))
 
-                    item = title, download_url, size, seeders, leechers
+                    item = {'title': title, 'link': download_url, 'size': size, 'seeders': seeders,
+                            'leechers': leechers, 'hash': ''}
+
                     if mode != 'RSS':
-                        sickrage.srCore.srLogger.debug("Found result: %s " % title)
+                        sickrage.srCore.srLogger.debug("Found result: {}".format(title))
 
-                    items[mode].append(item)
+                    results.append(item)
 
-            # For each search mode sort all the items by seeders if available
-            items[mode].sort(key=lambda tup: tup[3], reverse=True)
-
-            results += items[mode]
+        # Sort all the items by seeders if available
+        results.sort(key=lambda k: try_int(k.get('seeders', 0)), reverse=True)
 
         return results