+++ sickrage/providers/torrent/horriblesubs.py
@@ -20,7 +20,7 @@ from __future__ import unicode_literals
 
 import sickrage
 from sickrage.core.caches.tv_cache import TVCache
-from sickrage.core.helpers import bs4_parser
+from sickrage.core.helpers import bs4_parser, try_int
 from sickrage.providers import TorrentProvider
 
 
@@ -48,8 +48,6 @@ class HorribleSubsProvider(TorrentProvider):
         search_params = {
             "nextid": 0
         }
-
-        items = {'Season': [], 'Episode': [], 'RSS': []}
 
         for mode in search_strings.keys():
             sickrage.srCore.srLogger.debug("Search Mode: %s" % mode)
@@ -102,14 +100,15 @@ class HorribleSubsProvider(TorrentProvider):
                         if not all([title, download_url]):
                             continue
 
-                        item = title, download_url, size, seeders, leechers
+                        item = {'title': title, 'link': download_url, 'size': size, 'seeders': seeders,
+                                'leechers': leechers, 'hash': ''}
+
                         if mode != 'RSS':
-                            sickrage.srCore.srLogger.debug("Found result: %s " % title)
+                            sickrage.srCore.srLogger.debug("Found result: {}".format(title))
 
-                        items[mode].append(item)
+                        results.append(item)
 
-            # For each search mode sort all the items by seeders if available
-            items[mode].sort(key=lambda tup: tup[3], reverse=True)
-            results += items[mode]
+        # Sort all the items by seeders if available
+        results.sort(key=lambda k: try_int(k.get('seeders', 0)), reverse=True)
 
         return results