+++ sickrage/providers/torrent/scc.py
@@ -24,7 +24,7 @@ import urllib
 
 import sickrage
 from sickrage.core.caches.tv_cache import TVCache
-from sickrage.core.helpers import bs4_parser, convert_size
+from sickrage.core.helpers import bs4_parser, convert_size, try_int
 from sickrage.providers import TorrentProvider
 
 
@@ -80,13 +80,10 @@ class SCCProvider(TorrentProvider):
         return re.search(title, text, re.IGNORECASE)
 
     def search(self, search_strings, search_mode='eponly', epcount=0, age=0, epObj=None):
-
         results = []
 
         if not self.login():
             return results
-
-        items = {'Season': [], 'Episode': [], 'RSS': []}
 
         for mode in search_strings.keys():
             if mode != 'RSS':
@@ -121,7 +118,8 @@ class SCCProvider(TorrentProvider):
 
                             title = link.string
                             if re.search(r'\.\.\.', title):
-                                data = sickrage.srCore.srWebSession.get(self.urls['base_url'] + "/" + link['href'], cache=False).text
+                                data = sickrage.srCore.srWebSession.get(self.urls['base_url'] + "/" + link['href'],
+                                                                        cache=False).text
                                 with bs4_parser(data) as details_html:
                                     title = re.search('(?<=").+(?<!")', details_html.title.string).group(0)
                             download_url = self.urls['download'] % url['href']
@@ -142,16 +140,16 @@ class SCCProvider(TorrentProvider):
                                         title, seeders, leechers))
                             continue
 
-                        item = title, download_url, size, seeders, leechers
+                        item = {'title': title, 'link': download_url, 'size': size, 'seeders': seeders,
+                                'leechers': leechers, 'hash': ''}
+
                         if mode != 'RSS':
-                            sickrage.srCore.srLogger.debug("Found result: %s " % title)
+                            sickrage.srCore.srLogger.debug("Found result: {}".format(title))
 
-                        items[mode].append(item)
+                        results.append(item)
 
-            # For each search mode sort all the items by seeders if available
-            items[mode].sort(key=lambda tup: tup[3], reverse=True)
-
-            results += items[mode]
+        # Sort all the items by seeders if available
+        results.sort(key=lambda k: try_int(k.get('seeders', 0)), reverse=True)
 
         return results