+++ sickrage/providers/torrent/tntvillage.py
@@ -25,7 +25,7 @@ import sickrage
 from sickrage.core.caches.tv_cache import TVCache
 from sickrage.core.common import Quality
 from sickrage.core.exceptions import AuthException
-from sickrage.core.helpers import bs4_parser
+from sickrage.core.helpers import bs4_parser, try_int
 from sickrage.core.nameparser import InvalidNameException, InvalidShowException, \
     NameParser
 from sickrage.providers import TorrentProvider
@@ -280,9 +280,7 @@ class TNTVillageProvider(TorrentProvider):
                 if x['doc']['season'] == parse_result.season_number]) == len(parse_result.episode_numbers): return True
 
     def search(self, search_params, search_mode='eponly', epcount=0, age=0, epObj=None):
-
         results = []
-        items = {'Season': [], 'Episode': [], 'RSS': []}
 
         self.categories = "cat=" + str(self.cat)
 
@@ -400,20 +398,18 @@ class TNTVillageProvider(TorrentProvider):
                                                 title, seeders, leechers))
                                     continue
 
-                                item = title, download_url, size, seeders, leechers
+                                item = {'title': title, 'link': download_url, 'size': size, 'seeders': seeders,
+                                        'leechers': leechers, 'hash': ''}
+
                                 if mode != 'RSS':
-                                    sickrage.srCore.srLogger.debug("Found result: %s " % title)
+                                    sickrage.srCore.srLogger.debug("Found result: {}".format(title))
 
-                                items[mode].append(item)
-
+                                results.append(item)
                     except Exception:
-                        sickrage.srCore.srLogger.error(
-                            "Failed parsing provider. Traceback: %s" % traceback.format_exc())
+                        sickrage.srCore.srLogger.error("Failed parsing provider.")
 
-                # For each search mode sort all the items by seeders if available if available
-                items[mode].sort(key=lambda tup: tup[3], reverse=True)
-
-                results += items[mode]
+                        # Sort all the items by seeders if available
+        results.sort(key=lambda k: try_int(k.get('seeders', 0)), reverse=True)
 
         return results